Refactor equipment form
Implement the refactoring plan for the equipment form, creating new components and hooks to improve maintainability and organization.
