+++ src/components/equipment/form/EquipmentStatusLocationSection.tsx
@@ -1,0 +1,103 @@
+import React from 'react';
+import { Card, CardContent } from "@/components/ui/card";
+import { FormControl, FormField, FormItem, FormLabel, FormMessage } from "@/components/ui/form";
+import { Input } from "@/components/ui/input";
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
+import { UseFormReturn } from 'react-hook-form';
+
+interface EquipmentStatusLocationSectionProps {
+  form: UseFormReturn<any>;
+}
+
+const EquipmentStatusLocationSection: React.FC<EquipmentStatusLocationSectionProps> = ({ form }) => {
+  return (
+    <Card>
+      <CardContent className="pt-4 space-y-4">
+        <h3 className="font-medium text-sm text-muted-foreground uppercase tracking-wide">
+          Status & Location
+        </h3>
+        
+        <FormField
+          control={form.control}
+          name="status"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Status *</FormLabel>
+              <Select onValueChange={field.onChange} defaultValue={field.value}>
+                <FormControl>
+                  <SelectTrigger>
+                    <SelectValue placeholder="Select status" />
+                  </SelectTrigger>
+                </FormControl>
+                <SelectContent>
+                  <SelectItem value="active">Active</SelectItem>
+                  <SelectItem value="maintenance">Under Maintenance</SelectItem>
+                  <SelectItem value="inactive">Inactive</SelectItem>
+                </SelectContent>
+              </Select>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="location"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Location *</FormLabel>
+              <FormControl>
+                <Input placeholder="e.g., Warehouse A, Section 1" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="installation_date"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Installation Date</FormLabel>
+              <FormControl>
+                <Input type="date" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="warranty_expiration"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Warranty Expiration</FormLabel>
+              <FormControl>
+                <Input type="date" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="last_maintenance"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Last Maintenance</FormLabel>
+              <FormControl>
+                <Input type="date" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+      </CardContent>
+    </Card>
+  );
+};
+
+export default EquipmentStatusLocationSection;
